Implemented Delete a group by ID controller:
This commit adds a controller function for deleting a group by its ID.
It first attempts to find the group using findByPk, and if the group exists, it proceeds to delete it from the database using the destroy method.
After a successful deletion, it responds with a status code of 204 (No Content) and a success message indicating that the group has been deleted.
If the group is not found, it returns a 404 error indicating that the group was not found.
Error handling is included to handle any potential errors during the operation.

Also Added Middleware that check if user is logged in but commented it out for now, until login endpoint is implemented.